refactor: move error types and helpers to internal/apiclient/errors.go
Separate HTTPError type and extractAgentAuthError function into their
own file for better code organization, keeping them in the same
apiclient package.
